requirements.txt
SQLAlchemy==2.0.42
sqlmodel==0.0.24
alembic==1.16.4
fastapi==0.116.1
pydantic==2.11.7
pydantic-settings==2.10.1
bcrypt==4.0.1
psycopg2-binary==2.9.10
pyjwt==2.10.1
ruff==0.12.7
black==25.1.0
